Refactor movie and TV show headers to use Link component for homepages, remove UserMovieReviewList component and integrate its functionality into UserMovieReviewPage
